+++ packages/dashboard/CHANGELOG.md
@@ -1,5 +1,12 @@
 # argusai-dashboard
 
+## 0.15.1
+
+### Patch Changes
+
+- Updated dependencies
+  - argusai-core@0.15.1
+
 ## 0.15.0
 
 ### Patch Changes